Debug and finished

+++ routes/records.js
@@ -5,11 +5,13 @@ const repository = new Repository();
 const Record = require("../models/record");
 const HTTP_STATUSES = require("../models/httpStatuses");
 const HTTP_STATUS_MSG = require("../models/httpStatusMessages");
+const bodyParser = require("body-parser");
+const urlencoded = bodyParser.urlencoded();
 const jsonParser = express.json();
 
 router.get("/", (req, res) => {
     try {
-        return res.json(repository.getRecords());
+        return res.render("index.hbs", {records: repository.getRecords()});
     } catch (error) {
         return res.status(HTTP_STATUSES.INTERNAL_SERVER_ERROR).send(HTTP_STATUS_MSG.INTERNAL_SERVER_ERROR_MSG);
     }
@@ -18,6 +20,12 @@ router.get("/", (req, res) => {
 router.get("/:id", (req, res) => {
     try {
 
+        if (!req.params) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+        if (!req.params.id) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+
+        const id = +req.params.id;
+        if (!id || id <= 0) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+
         return res.json(repository.getRecord(req.params.id));
 
     } catch (error) {
@@ -25,12 +33,14 @@ router.get("/:id", (req, res) => {
     }
 });
 
-router.post("/", jsonParser, (req, res) => {
+router.post("/", urlencoded, (req, res) => {
     try {
 
+        if (!req.body) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+
         const lastRecordId = repository.getLastRecordId();
-        const title = req.body?.title;
-        const description = req.body?.description;
+        const title = req.body.title;
+        const description = req.body.description;
 
         if (!title) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
         if (!description) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);;
@@ -38,20 +48,24 @@ router.post("/", jsonParser, (req, res) => {
         const record = new Record(lastRecordId + 1, title, description);
         repository.addRecord(record);
 
-        return res.json(record);
+        return res.render("index.hbs", {records: repository.getRecords()});
 
     } catch (error) {
         return res.status(HTTP_STATUSES.INTERNAL_SERVER_ERROR).send(HTTP_STATUS_MSG.INTERNAL_SERVER_ERROR_MSG);
     }
 });
 
-router.put("/:id", jsonParser, (req, res) => {
+router.put("/:id", urlencoded, (req, res) => {
     try {
-        const id = req.params.id;
-        const title = req.body?.title;
-        const description = req.body?.description;
 
-        if (!id) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+        if (!req.params || !req.params.id) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+        if (!req.body) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+
+        const id = +req.params.id;
+        const title = req.body.title;
+        const description = req.body.description;
+
+        if (!id || id <= 0) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
         if (!title) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
         if (!description) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
 
@@ -61,7 +75,7 @@ router.put("/:id", jsonParser, (req, res) => {
         const record = new Record(id, title, description);
         repository.updateRecord(id, record);
 
-        return res.json(record);
+        return res.render("index.hbs", {records: repository.getRecords()});
 
     } catch (error) {
         return res.status(HTTP_STATUSES.INTERNAL_SERVER_ERROR).send(HTTP_STATUS_MSG.INTERNAL_SERVER_ERROR_MSG);
@@ -71,14 +85,18 @@ router.put("/:id", jsonParser, (req, res) => {
 router.delete("/:id", (req, res) => {
     try {
 
-        const id = req.params.id;
-        if (!id) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+        if (!req.params) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
+
+        const id = +req.params.id;
+        if (!id || id <= 0) return res.status(HTTP_STATUSES.BAD_REQUEST).send(HTTP_STATUS_MSG.BAD_REQUEST_MSG);
 
         repository.deleteRecord(id);
 
-        return res.status(HTTP_STATUSES.NO_CONTENT).send();
+        return res.render("index.hbs", {records: repository.getRecords()});
         
     } catch (error) {
         return res.status(HTTP_STATUSES.INTERNAL_SERVER_ERROR).send(HTTP_STATUS_MSG.INTERNAL_SERVER_ERROR_MSG);
     }
 });
+
+module.exports = router;
